+++ js-miniapp-bridge/src/types/log-type.ts
@@ -1,0 +1,19 @@
+/**
+ * Enum representing the types of logs.
+ */
+export enum LogType {
+  /**
+   * Debug log type.
+   */
+  DEBUG = 'debug',
+
+  /**
+   * Info log type.
+   */
+  INFO = 'info',
+
+  /**
+   * Error log type.
+   */
+  ERROR = 'error',
+}